+++ packages/editor/src/hooks/index.js
@@ -3,6 +3,5 @@ import './default-autocompleters';
 import './media-upload';
 import './pattern-overrides';
 import './navigation-link-view-button';
-import './template-part-navigation-edit-button';
 import './push-changes-to-global-styles';
 import './note-format';